room: move flipmap logic to TRX
This moves the main flipmap logic itself to TRX.
