+++ src/libtrx/include/libtrx/game/sound/common.h
@@ -18,6 +18,7 @@ void Sound_ResetSources(void);
 void Sound_PauseAll(void);
 void Sound_UnpauseAll(void);
 
+extern void Sound_StopAmbientSounds(void);
 extern void Sound_StopAll(void);
 extern bool Sound_Effect(
     SOUND_EFFECT_ID sfx_num, const XYZ_32 *pos, uint32_t flags);
