+++ src/tr2/game/room.c
@@ -744,88 +744,6 @@ void Room_AlterFloorHeight(const ITEM *const item, const int32_t height)
     }
 }
 
-void Room_FlipMap(void)
-{
-    for (int32_t i = 0; i < Room_GetCount(); i++) {
-        ROOM *const room = Room_Get(i);
-        if (room->flipped_room == NO_ROOM_NEG) {
-            continue;
-        }
-
-        Room_RemoveFlipItems(room);
-
-        ROOM *const flipped = Room_Get(room->flipped_room);
-        const ROOM temp = *room;
-        *room = *flipped;
-        *flipped = temp;
-
-        room->flipped_room = flipped->flipped_room;
-        flipped->flipped_room = NO_ROOM_NEG;
-        room->flip_status = RFS_UNFLIPPED;
-        flipped->flip_status = RFS_FLIPPED;
-
-        // TODO: is this really necessary given the assignments above?
-        room->item_num = flipped->item_num;
-        room->effect_num = flipped->effect_num;
-
-        Room_AddFlipItems(room);
-    }
-
-    Room_ToggleFlipStatus();
-}
-
-void Room_RemoveFlipItems(const ROOM *const room)
-{
-    int16_t item_num = room->item_num;
-
-    while (item_num != NO_ITEM) {
-        ITEM *const item = Item_Get(item_num);
-
-        switch (item->object_id) {
-        case O_MOVABLE_BLOCK_1:
-        case O_MOVABLE_BLOCK_2:
-        case O_MOVABLE_BLOCK_3:
-        case O_MOVABLE_BLOCK_4:
-            Room_AlterFloorHeight(item, WALL_L);
-            break;
-
-        default:
-            break;
-        }
-
-        if (item->flags & IF_ONE_SHOT
-            && Object_Get(item->object_id)->intelligent
-            && item->hit_points <= 0) {
-            Item_RemoveDrawn(item_num);
-            item->flags |= IF_KILLED;
-        }
-
-        item_num = item->next_item;
-    }
-}
-
-void Room_AddFlipItems(const ROOM *const room)
-{
-    int16_t item_num = room->item_num;
-    while (item_num != NO_ITEM) {
-        const ITEM *const item = Item_Get(item_num);
-
-        switch (item->object_id) {
-        case O_MOVABLE_BLOCK_1:
-        case O_MOVABLE_BLOCK_2:
-        case O_MOVABLE_BLOCK_3:
-        case O_MOVABLE_BLOCK_4:
-            Room_AlterFloorHeight(item, -WALL_L);
-            break;
-
-        default:
-            break;
-        }
-
-        item_num = item->next_item;
-    }
-}
-
 void Room_Legacy_TriggerMusicTrack(
     const int16_t track, const int16_t flags, const int16_t type)
 {
